app/requirements.txt
flask
flask-reggie
mysql-connector-python>=8.0